Refactored the Form 2
