+++ frontend/src/components/PregnancyPartB/Pregnancy.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import Tables from './TablesR1';
 import Table2 from "./TableR2";
 import Pregnancy2 from "./Pregnancy2";
-import { useForm2Context } from './contexts/Form2Context';
+import { useForm2Context } from '../../contexts/Form2Context';
 
 const registrationFields = [
     { id: 'regNo', label: 'Registration No.:', type: 'text' },
