+++ frontend/src/components/PregnancyPartB/TableR2.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useForm2Context } from './contexts/Form2Context';
+import { useForm2Context } from '../../contexts/Form2Context';
 
 const Table2 = () => {
     const { formData, setFormData } = useForm2Context();
